server/requirements.txt
python==3.11
SQLAlchemy==2.0.0			# SQLAlchemy for ORM
psycopg2==2.9.3		        # PostgreSQL adapter for SQLAlchemy
pytest==7.2.2				# Pytest for testing
pyzmq==25.1.1				# PyZMQ for ZeroMQ bindings in Python
python-dotenv==1.0.1        